Create default user for "master" org too
Summary:
Creates the default user for the master org too.

Note -- I removed a migration and added the field to creating the organization.
The reason is because the value was null, and handling a new migration would be
problematic.  This assumes everyone who needs the mutated organization already
has it, and this is only for new installations

Reviewed By: naor9991

Differential Revision: D16185580

fbshipit-source-id: a9c1f5324a085f53a74c4191138bf3ce83e58532