Async-ify analytics components
Summary:
The analytics container in the NMS backend currently processes jobs synchronously. This is the first of a series of diffs (which I intend to land together once they're all ready) that restructures the inner workings of the framework to execute asynchronously.

I've adopted the Producer-Consumer model detailed [here](https://asyncio.readthedocs.io/en/latest/producer_consumer.html) in the asyncio developer docs. In this instance, the producers will read the pipeline configuration from `PipelineConfig.json` and create and add jobs (whose definitions will reside in `insights.py`) onto the shared queue. Consumers pop jobs off of the queue and then execute them to completion.

A pipeline configuration is as follows:
- jobs: a list of functions to add to the queue and run
- period: how often to add the jobs from this pipeline to the queue
- step: query resolution step width
- params: dict with custom params per job

e.g.

```
  "network_watchdog": {
    "period": 30,
    "jobs": [
      {
        "name": "detect_ignition_failure",
        "enabled": false,
        "params": {
          "link_white_list": [],
          "link_black_list": []
        }
      }
    ]
  }
```

In order to make this possible asynchronously, I've created the following base classes:

- http_client.py: for interfacing with external HTTP endpoints
- mysql_connection_pool.py: for interfacing with a MySQL database
- prometheus.py: for reading stats from a Prometheus database

As a small formatting nit, I've changed the order of the requires in `setup.py` to be alphabetical and switched to using `==` instead of `>=` to help us debug in the future

Reviewed By: satyakb

Differential Revision: D16041216

fbshipit-source-id: 5e312d2fee6e4871d9b73800f13f92af03d33b1c