fix Ref forwarding crashes
Summary:
Core breaking change:
https://material-ui.com/guides/migration-v3#core
"Every component forward their ref. This is implemented by using React.forwardRef(). This affects the internal component tree and display name and therefore might break shallow or snapshot tests. innerRef will no longer return a ref to the instance (or nothing if the inner component is a function component) but a ref to its root component. The corresponding API docs list the root component."

ref: This is a react feature. Use refs to target a dom element

fwdRef: This is an FBC convention backed by flow and a helper. Sometimes a ref needs to be passed through an intermediary component down to the nearest dom node in the tree.

innerRef: this is a convention used by some libraries to target the _instance_ of a child react component. This is fundamentally different from fwdRef and should only be used if the library requires it.

Reviewed By: tcirstea

Differential Revision: D16292482

fbshipit-source-id: 612bac0ada7de315e24c24d71284080cd4668380