nms/beringei: add rdkafka/cppkafka to build
Summary:
Adding kafka support requires us to bump the zstd version to keep it supported with rdkafka, which required building the libraries with the system versions instead of the versions we compile in /usr/local/facebook*.
This will be easier to deal with once beringei code is gone, since it'll only be our code that we need to bring up support for with dependencies.

Also deleting cmake temp files and all of the library code - which keeps the shared objects to link against.

Reviewed By: elludraon

Differential Revision: D16531213

fbshipit-source-id: e56585c135abaeef3e3d63e263c2b65d0c7a9030