Add service availabilty stats to overview panel
Summary:
Service availability is loosely defined as a measure of availability to the subscribers of the network. We use the UDP pinger loss ratio data which to compute this metric. The pinger pings each node on the network from wherever the NMS backend is installed counts the number of packets that don't return from each ping.

Eventually, we would like to provide a bandwidth with each availability number. Displaying service availability is disabled by default, but can be enabled by changing the `SERVICE_AVAILABILITY_ENABLED` environment variable to true.

Reviewed By: aclave1

Differential Revision: D16441014

fbshipit-source-id: 41d4a716a9a82c75ad788f653c200dec4bc9c345