better sqlite mocking
Summary: In D16656512 i built a "mockDatabase" function which handled the sequelize config mocking. While this handled the case where the test file is importing models directly (for seeding) it failed the case where the file being tested imported from models (metrics.js). This is all due to the way that es imports work and how they're hoisted to the top. We could have mandated using `require()` calls instead and relied on ordering of requires, but this is brittle and relies on undocumented side-effects. This new solution handles both cases almost transparently but still delegates to the real underlying models once they've been configured properly.

Reviewed By: omikader

Differential Revision: D16669033

fbshipit-source-id: 7053a8ddc8e5a43bde5b0864dd51834c201b9cfb