Fix offline whitelist bug
Summary:
The problem was in my usage of lodash's difference function. When I decide whether to save the new whitelist or not, I use lodash's difference function to determine if any new keys have been added to the whitelist. Lodash's difference function states:

> The order and references of result values are determined by the first array

Since I was doing `_.difference(old, new)` it was enumerating keys of old and looking to see if any of them were not in new. This means that the whitelist would effectively never be updated after the first save. The correct way is to just swap the order and check if new has any keys that are not in old.

Reviewed By: elludraon

Differential Revision: D16692785

fbshipit-source-id: 75d82877c06d843422c3bd99b08ec052b1d20181