Enable channel features in M42 instead of M43
Summary: I misunderstood how the release numbering worked and thought we were in M43 and not M42 last week. Changed the minimum version for the "Optimize Channel Allocation" button to M42 since that's the controller release the feature in the controller shipped in. [Link to relevant release notes](https://tgdocs.internal.tfbnw.net/releases.html#patrick-wolfe-pwolfe).

Reviewed By: elludraon

Differential Revision: D16834889

fbshipit-source-id: 4600445fa2354abb58e8f8cb2efe1c670e89f727