Overview panel tests
Summary:
Adds some basic tests for the overview panel. The main purpose of this is to create examples for testing the map panels. They require a lot of props and are tightly coupled to context and topology config.

Note:

This also adds some(experimental) topology helpers to add nodes and links to a topology with less effort. I'm not sure how useful these are when operating only on the topology. I think it'd be better if they operated on NetworkContext so we can keep the siteMap, nodeMap, and linkMap objects in sync with the mock topology. I will likely add this in a new diff once its actually required.

Differential Revision: D16858020

fbshipit-source-id: 62648c46830942fc5e636133a1f8b6ab0f98d699