login race condition
Summary: User's authentication token is stored in the express session. The session is backed by sequelize. Whenever req.logIn is called, the session is _mutated_ by passport, and we redirect in the `done` callback. Instead of saving the session explicitly, passport relies on express-session's change detection mechanism. Express-session detects that the session object has changed and then saves it at the end of the request lifecycle. This can introduce a race condition if a redirect is immediately issued inside of the done callback. This fixes the race condition by explicitly calling the save function.

Differential Revision: D16940747

fbshipit-source-id: a072d0cb5ed8d66a217ee0eaee4ae4da97d83b9d