Move tglib framework into separate folder
Summary: This diff moves all the shared microservice code that I introduced in D16041216 into a separate folder in the `connectivity-labs` repository. In addition, I've added several other clients for communicating with Kafka and API service. I am still iterating on this diff and I welcome any suggestions in ways to improve the framework, whether it be with regards to code structure or usability or anything in between. Please start patching these changes and using them so we can begin to productionize.

Reviewed By: csmodlin

Differential Revision: D16726247

fbshipit-source-id: b47eadfa0ea031dd83cc952640f9bdd28b8e2812