Only make core-js not hoisted for relay-compiler
Summary:
We have an issue where node_modules isn't persisted when we mount docker volumes,
so everything breaks.  We haven't seen this until now, because we have just
started using `nohoist`.  We don't have a solution for this, but this commit works
around it by isolating corejs 2 to *just* `fbcnms/relay` (which is the only place
that needs it).  Luckily for us, that is also only run on the host machine.

Reviewed By: Scott8440

Differential Revision: D17085173

fbshipit-source-id: ddbae782a7aa178541a8c5c5cfcb1cd97899063a