Move base cxl models to tglib
Summary: Although it is a goal for all microservices to own and operate their own databases, it is not something that we can enforce for existing services such as network test and the scan service right away. This is because many of their tables are already defined in `cxl`. As a result, this diff defines the schemas of the base tables that they interact with such as `controller` and `topology`.

Reviewed By: noahfren, csmodlin

Differential Revision: D17106073

fbshipit-source-id: 1544cf1600495fa2f98b573ce6019add3135db7e