Add Dockerfile for tglib
Summary:
The plan is for all new microservices to use this base image. The base image comes preloaded with glib installed, along with python3.7 (which is symlinked to python).

In addition, I've created and installed a new module called `tgif` which houses all of the Terragraph specific thrift source and generated python files.

Reviewed By: satyakb

Differential Revision: D17066271

fbshipit-source-id: e9d3da9e463da5ec6d86dfe22a928459ee9dbe9c