Merge EventClient with KafkaProducer
Summary:
EventClient was originally designed to be a standalone client for `tglib` developers to use to log events with the purpose of adding client-side auditing to our first-class services. Now that this is no longer our direction, I've combined the functionality (3 specific event logging functions) with that of KafkaProducer as they no longer need to be distinct components.

Under the hood, both are inheriting from BaseClient and AIOKafkaProducer, and both have the same start/stop logic, so this was a logical next step.

Reviewed By: csmodlin

Differential Revision: D17132745

fbshipit-source-id: 85e7a2a44e2a0705d9faf6d1831d1ad51fc9f1f2