Rename tgif folder to if
Summary: Compiling thrift into the `tgif` package did not work due to conflicting namespaces when importing thrift files in one another. As a result, I've renamed `tgif` to be `if` as it will only house the raw thrift files. I've added a `.keep` file in order to make sure that the empty directory is actually committed.

Reviewed By: elludraon

Differential Revision: D17239270

fbshipit-source-id: 8d47d806404147a5862f788a78817f6af063a9a4