self signed cert for keycloak in docker-compose, copy keycloak files
Summary:
aclave1 added a support script for setting up keycloak with our realm settings in D17189027
This adds a self-signed cert with our CA (but the CA isn't needed yet, we'll add that later) and copies the necessary files.

I tried integrating the scripts to run with ansible, but failed so that will be in the next round.

Reviewed By: aclave1

Differential Revision: D17239404

fbshipit-source-id: 928a70196a5f597f1bb107cbc9c65c589d8f1969