mysql link availability via kafka stream [WIP]
Summary:
Add support for reading the DB directly for link events. This is gated on env.STATS_BACKEND = prometheus (which means kafka/prometheus now..)
QS reads the kafka stream and writes the link uptime start/end times, now we can keep this data forever.
When a new time window is selected we immediately query the DB for the period. 24 hours is still the default cached value.

TODO:
- handle end of time window - I'm planning to give a grace period at the end so we still assume it's online if the last ~2 minutes are missing.
- should we restrict the viewing to the earliest occurrence of DB data so the availability #s aren't mis-stated?

Reviewed By: aclave1

Differential Revision: D15921657

fbshipit-source-id: 6bff334392edb103af928abbba18a13816751738