Change client design pattern
Summary:
The Singleton design pattern is not very Pythonic, and after discussing with Noah while he was working on the MongoDB client, we arrived at a better design pattern (where each underlying resource is a class variable shared among all the instances). This is simpler to follow in the event that we add more clients in the future, and it conforms better to common Python practices.

All tests and examples were updated to reflect these changes.

Reviewed By: noahfren

Differential Revision: D17352883

fbshipit-source-id: ea3b8177d4f8625d6d5122a1b74183ee799bf17e