fix NMS failing to start when database is down
Summary:
When the mysql instance is down while NMS is trying to start up, NMS will fail to start, even when the database does come back up. If the database goes down after NMS has successfully initialized, there will be no problem.

The fix was to loop at startup until the NMS can successfully connect to the database. This will still prevent NMS from connecting to its port and serving web traffic, but it will properly start once it's able to connect to the db instance

Reviewed By: omikader

Differential Revision: D17501548

fbshipit-source-id: ebc7b2e1ccddb4e43fed30dba506f16f878c257b