Don't add unsuccessfully started clients to shutdown list
Summary:
There can be failures when the clients are started sometimes. When this happens, the ones that were started successfully should be cleanly shut down and the failure should be raised properly.

Previously, all clients were being shut down, including ones that didn't successfully start. This caused a problem where `ClientStoppedErrors` were being raised because the clients never started to begin with. This diff changes is so that only clients who's `start` results don't belong to the family of `TGLibErrors` are added to the stop list.

Reviewed By: noahfren

Differential Revision: D17514506

fbshipit-source-id: c3fd70b36097a0b77247f77f26bce8e84682f39c