Add ability to return exceptions in APIService client
Summary:
It is possible that not all API service instances are running when a `request_all` or `request_many` request comes in from a tglib microservice. When this happens, the current behavior is for the microservice to propagate the exception indicating that an API service is unavailable.

On certain occasions, (getTopology), we like to give service developers the ability proceed with execution and process/ignore the error based on their own implementation logic. This diff adds the ability for the user to do as much by passing in a boolean flag which is forwarded to the `asyncio.gather` call.

Reviewed By: spurav

Differential Revision: D17537871

fbshipit-source-id: e273a77cb70fa88f1804d0766d33586ed012e3f7