Add UDP Pinger dashboards as default
Summary:
We would like UDP pinger dashboards to persist in between Grafana upgrades/restarts. The way we do this is by exporting the dashboard JSON and importing it using the grafana utility container at creation time.

Node and Network panels are separated due to long load times for templating node variables.

Hardcoded `uid` values are supplied in the event that we would like to link to the dashboards directly from NMS.

Reviewed By: csmodlin

Differential Revision: D17460216

fbshipit-source-id: fa307b3220b3e9f29690797b1b6be60189b24d51