Combine network and node dashboards
Summary: These dashboards were initially separated because loading the node variables took an extremely long amount of time. Cory noticed that the node name source query could be made to be more efficient and load the node variables in a much shorter amount of time! This diff combines the dashboards using this technique.

Reviewed By: csmodlin

Differential Revision: D17667305

fbshipit-source-id: 0a215a0a46f0f83ed5820217d9145277106beeca