Redefine pinger rate
Summary:
The NetNorad implementation of the UDP pinger calculates the pings per second (pps) by dividing the number of hosts (targets) by the time it took to complete the sweep.

For our purposes, we want the `pinger_rate` to be representative of the the amount of time to wait in between each sweep across ALL hosts. i.e. `pinger_rate = 5` would mean that we should send at a rate of 5 pps so there should be 200ms spacing in between each sweep.

Reviewed By: pmccut

Differential Revision: D17724114

fbshipit-source-id: 2713902201da1c0198a2e2bb596cd773db1c9dcd