Full Upgrade button
Summary: In NMS V1 there was a full upgrade button, but the upgrade was not super stable. We are at a point now where upgrades should be pretty automatic and succeed most of the time without any modifications. So we are bringing back Full Upgrade for users to save time when upgrading networks.

Reviewed By: aclave1

Differential Revision: D17791779

fbshipit-source-id: cf8a11793ad1ae9fe42955f4f8db2dd0cc9aabd4