fix networktest crash
Summary:
In mpk production, the network test help api does not return the asap parameter. This is likely because the networktest container is outdated and has none of the scheduling functionality. This code patches around that and also adds some tests that return the incorrect api responses to ensure it doesn't break again.

In the future, we should add tests which also return the correct api responses.

Differential Revision: D17913476

fbshipit-source-id: 9111b9bb9892a82f842e701e571f445467dc2618