Make cookie timeout configurable
Summary:
Users are tired of logging in every day when they want to use NMS.

The cookie timeout can now be configured via the SESSION_MAX_AGE_MS environment variable.

If this variable is not set, and LOGIN_ENABLED is set, the expiration is far in the future. The user's keycloak access token will expire long before this, necessitating another signin. This effectively allows the user to configure their signin duration via the keycloak ui or an environment variable.

Reviewed By: rckclmbr, elludraon

Differential Revision: D17953480

fbshipit-source-id: 17b798e3a4cb809d34a8f8a847b3a33abd68946a