serve NMS traffic sooner, add health check endpoint
Summary:
This diff changes the order so that NMS can serve traffic sooner and adds a healthcheck endpoint which requires no auth.

If login is enabled and NMS cannot access keycloak, NO traffic is served, NMS can't even serve an error page. This is because app.listen was invoked after some tasks which required keycloak.

This is a problem in ansible/swarm because there is a step that waits for NMS to begin serving traffic before continuing. If NMS is setup before keycloak, it cannot possibly access keycloak. Since ansible cannot continue and will fail the NMS task, keycloak can never start if it occurs after.

Differential Revision: D18045404

fbshipit-source-id: d0a0ffffa54a7dc59062de441f196787484c221a