fix broken docker build by pinning node:alpine version to 12
Summary:
The docker build started failing 2+ days ago. This seems to be due to a newer version of node:alpine, as they just switched from 12 -> 13.

We're failing to find a python version to install an optional dependency snappy for kafka. We need to fix that as well, but that was failing prior so we'll fix it separately.

Docker build failing (noticed in jenkins) - P120029789
Docker build with version from a few days ago - P120029585

Reviewed By: aclave1

Differential Revision: D18147111

fbshipit-source-id: b7f4ab3f15363248af09e401dbbcbf76bcfcf8ff