Update interface selection logic
Summary:
The UDP pinger was failing to run in MPK production because the API service requests were timing out. It was difficult to find out why this was happening because the entire CURL request was nested in a "catch all" exception.

- Increase CURL timeout to 5s
- Refactor CURL implementation in ApiServiceClient.h

Sometimes the globally scoped address was not being assigned to eth0 which caused the pinger to expect responses on the wrong address.

- Change the default source interface from "eth0" to ""
- Add logic to skip over addresses that match the default source interface name **but aren't** global scope

Reviewed By: elludraon

Differential Revision: D18239341

fbshipit-source-id: d26e0ca7896b5a3410f5a509fb3fd5a498637a87