adding force reload of page if axios request sent while logged out
Summary:
If an axios request is sent while user is logged out/not authenticated, res.redirect() doesn't work. Plus, it would redirect to the wrong url i.e. instead of the url of the page that sent the axios request, it would load the url of the request itself.

followed solution from https://github.com/axios/axios/issues/932#issuecomment-307390761 to distinguish between axios request and regular browser navigation requests. On the client side, we force reload the page since that will give us the correct login url.

question: where is a better location to put `axios.interceptors`??  It sets interceptor for axios globally so it only needs to be run once. I tried putting it in `axiosConfig.js` but that didn't work.

Reviewed By: rckclmbr

Differential Revision: D18364572

fbshipit-source-id: e5a8e75da7d7fe0e7b7cc036e7286a34661e4b73