Prevent generateAPIFromSwagger.sh from adding incorrect line
Summary:
`Using Yarn from /Users/alanw/Repos/fbsource/xplat/third-party/yarn/yarn` appears in the generated file when I run it.

Internal script was discovered to respond to `--silent`, but not `-s`. This diff updates to use the longer form flag which is properly interpreted by both the wrapper script and yarn itself.

Reviewed By: murtadha

Differential Revision: D18622486

fbshipit-source-id: d93f5e951cd14812478e49e06414fa7b06f0a60a