refresh query service status periodically
Summary:
The status of online/offline is piggy-backing on other calls going to query service. Now that we're hitting prometheus directly I'd rather have this periodic so it's always up to date. Also we were using the name beringei to refer to query service in too many places. Renamed so it's correct.

For later - we could show more services for easier troubleshooting - say prometheus status or how many stats prometheus is processing.

Reviewed By: aclave1

Differential Revision: D18715118

fbshipit-source-id: f4a9dafb6d459b805648539178deea36272d2da7