Add tests for getTestResults in NetworkTestAPIUtil
Summary: I only added tests for this one function because network test is being rewritten. A lot of what we'd test for is fixing little quirks with the API but since the api is changing, there will be different quirks. I finished this up because I wanted us to have some kind of base to work with once the new api is done

Differential Revision: D18783669

fbshipit-source-id: 624feb73d710db4ae5df6acdeea37525e8f88722