allow delay in link health availability
Summary: The link health % calculations need to account for a slight lag in stats data coming from the nodes into kafka and being processed into link health events. This allows a 2 minute delay for this process - removing up to 2 minutes if the last interval is LINKUP. Once the window is exceeded we assume the event is accurate and the +2 minutes allowance vanishes. 2 minutes picked because it's worked so far but we might need to adjust this later.

Reviewed By: aclave1

Differential Revision: D18797791

fbshipit-source-id: 82f11cdf187a2c022e2a779f7fed75dcfc26afe5