Use fbthrift instead of apache thrift
Summary:
Sandcastle uses fbthrift when compiling the thrift files needed for tglib. This was causing conflicting results in ptr because it was using the Apache thrift python library as prescribed by the tglib setup.py file.

The "easiest" fix is to change tglib to use fbthrift as well. I've installed all of the necessary dependencies in order to install fbthrift.

Reviewed By: satyakb

Differential Revision: D18823675

fbshipit-source-id: bbae5edde85fcc16d9ba605d395e4f9ab60a2ff5