support prometheus cache service
Summary: We have been using our own queue for storing metrics for prometheus to scrape. We separated this into a 1s and 30s queue. Scott8440's service does the same thing, so we can publish the metrics to this service via curl and use it as our queue. A single queue for all metrics should be equivalent - since we don't add all of the metrics to the queue at the same time we shouldn't see too much 30s data at the same time slowing down the scrape.

Reviewed By: omikader, Scott8440

Differential Revision: D19050750

fbshipit-source-id: f9f2aca0c5011c8dab3ba5f293b7f883a47e5faf