Add keycloak support
Summary:
The UDP ping client communicates with API Service every 60s in order to refresh its list of nodes to ping. Now that keycloak is deployed to manage access permissions to API Service and NMS, we must make sure that the pinger is keycloak compliant.

The new workflow is:

1) Every 60s, get a JWT from keycloak using the keycloak environment variables. By default, a JWT will expire after 300s but this value is configurable (as is the 60s refresh period). If the JWT has expired, fetch a new one.

2) Using the JWT, make getCtrlStatusDump and getTopology requests to create the UDP ping test plans.

3) As before, ping using the test plans.

While I was working in the codebase, I decided to unify the various curl implementations into one function that works for all of them. I also deleted unused code in PrometheusUtils.h/cpp

Reviewed By: pmccut

Differential Revision: D19061491

fbshipit-source-id: e016f3acfda870f4db8ab39ed403f9ff1c14ed53