Fixing Typeahead suggestions being hidden
Summary:
In case of a (position absolute\relative + overflow different of visible) parent container (of any level), if the Typeahead suggestions panel exceeds the parent's boundaries, it would be cut or not shown at all.
The change fixes that, yet it doesn't allow us to set the Typeahead suggestions panel's min-width to the input container.

Reviewed By: naor9991

Differential Revision: D19257371

fbshipit-source-id: 52abc1a76a2890b5b844531c3931a40995c88582