Select receiver for alert rule
Summary:
Allows the user to select a receiver for their alert rule to send alerts to. This works by creating a route in the route tree with an exact match for the alert name.

other changes:
* Changes the UI to match the latest UI designs
* Introduces RuleEditorBase, a component which handles layout and common editing tasks such as receiver, summary, and labels

UI after switching to jing's design
https://pxl.cl/W62R
https://pxl.cl/W62T

here's how add/edit receiver looks after moving the buttons
https://pxl.cl/W62W

Reviewed By: Scott8440

Differential Revision: D19188924

fbshipit-source-id: 097c487a055760b517e30272a32d9b8d6fb80d75