Update examples and small fixes
Summary: The example code was still referring to 'main' when it should now be referring to 'async_main'. Other fixes include eradicating the remaining usages of string formatting in favor of f-strings, having the prometheus endpoint return a JSON list instead of a string, and updating the endpoints for setting/updating service configuration.

Reviewed By: spurav

Differential Revision: D19642967

fbshipit-source-id: 803d2bc16c96dcc8a0f65627ee438a648441e460