Restart schedules after stopping errant sessions
Summary:
The current implementation tries to start all the schedules before stopping all of the running tests and marking them as ABORTED. It makes more sense for the schedules to be started after this step because the schedules could potentially (in very rare cases) start a test immediately.

I also removed the ptr run from the Dockerfile for network_test because it's not necessary to run this step during the build process. Sandcastle should catch errors and ptr should be used in the container as a development tool/resource, not at build time.

Reviewed By: spurav

Differential Revision: D19697494

fbshipit-source-id: 95e9107dc746b6ea87fcb5a39767e9a89766527c