test normal users can only see their networks (#1160)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1160

I wrote this test in an attempt to reproduce an error a partner was experiencing.  It turns out there's no bug in the code and the partner was simply on a very old revision, but we might as well have this test for the future to make sure this doesn't break.

Reviewed By: rckclmbr

Differential Revision: D19672542

fbshipit-source-id: 2df70ac0fcb7f4da1283372799e8a50078974635