Alarms env var updates
Summary:
The environment variables for integrating with the alarms service included hostnames and ports. This diff specifies the whole url like http://hostname:port in one single string and renames the env vars to better reflect this shift. The renamed variables were experimental and dev-only so there is no chance for breakage of existing setups.

* If using tgnms alarms, you'll need to update the vars in your .env file.

Reviewed By: pmccut

Differential Revision: D19725108

fbshipit-source-id: aa873002d89fd06619f6eb10c8b84f74ec82c372