Generalize keycloak service account client script
Summary:
NMS is configured in keycloak using the generate-keycloak-client script. Since the nms script is tested and working, i've decided to create a new script for new service account clients. I've added a deprecation comment to the old script. We can migrate NMS to the new script once it's tested with MSA.

This script should be run from within a jboss/keycloak container (it utilizes some java based utilities specific to keycloak). The script needs to be bind-mounted to the container and then run as the container's primary command.

Usage example:

```
docker run --rm
    --mount type=bind,source={{ keycloak_path  }}/docker_volumes/keycloak/scripts,target=/scripts \
    --env-file={{ keycloak_path }}/env/keycloak.env \
    --mount type=bind,source={{ msa_path }}/env/msa_keycloak.env,target=/output.env \
    --entrypoint="" \
    --network=terragraph_net \
    -u root \
    jboss/keycloak:7.0.0 \
    /scripts/create-service-account-client default_routes_service /output.env tg_topology_read,tg_management_read
```

* The first bind mount to /scripts is just to get the script into the container so that it can be run as the container's primary command (last line).
* The second bind mount is the script's output mount. This example mounts an existing .env file /msa_keycloak.env as /output.env inside the container. Then in the container's primary command, /output.env is passed as the script's output argument. If, for example, the msa_keycloak.env file didn't exist, you could just as easily do something like:

```
--mount type=bind,source={{ msa_path }}/env,target=/output \
/scripts/create-service-account-client service_name /output/env_file.env role_1
```

Reviewed By: omikader

Differential Revision: D20234322

fbshipit-source-id: 48aafbacd1022d3572efd90416c2eb05a4b87e11