Add ansible support for tglib-based microservices
Summary:
We want to start adding functional microservices to the our production setup (ansible). I added the default_routes_service and topology_service to the already existing `msa` role (I deleted the old network test because it doesn't work).

The main feature in this diff is that MySQL credentials are generated randomly on the ansible controller machine and used to create the user/database per MSA service. Since the alembic.ini file needs the credentials and cannot load them from the environment, I've templated it and deleted the hardcoded versions in the service directories.

The other change is that I added the MySQL root credentials to the group_vars section since keycloak and the msa roles both need them.

Reviewed By: satyakb

Differential Revision: D20142457

fbshipit-source-id: 942eb05c18c5f5f53725550964c1c4000488ad5d