Update DRS history endpoint
Summary:
DRS's `/history` api endpoint returned default route changes for any node in the given time window. If there were no route changes in that window, the service would return an empty result. This was misleading as the services implied that there were no default routes for the node in the given time window.

This diff updates `/history` api endpoint to fetch the latest default routes before the given time window, along with the route changes in that window.

Case 1:
If there are no routes for the node in the database, the service will return empty results.

Case 2:
If there are routes for the node in the time window, the service will return routes change history and the utilization for those routes in that window.

Case 3:
If there are routes for the node in the database, but none in the time window, the service will return the latest routes before the start time of the window, and the utilization will be 100%.

Reviewed By: omikader

Differential Revision: D20208286

fbshipit-source-id: dbdbba7816ff809707d85c0206dc12e8b14a4fb2