Change logging verbosity to 'exception' inside exception handlers
Summary:
It's better to use `logging.exception` in try-except blocks because we don't need to extract the exception message using "repr" or "str" and we get the entire exception trace. The latter can be particularly helpful when an exception is being "raised from" another exception and we want to know the root cause.

Also I renamed a couple of variables to be more accurate.

Reviewed By: hmehmood0

Differential Revision: D20353590

fbshipit-source-id: ef8f992007b6207ed49f0eafc31ef3050c66b78b