Fix issue where routes found no matching schedule/execution
Summary: The `list_schedules` and `list_executions` functions could return `None` if they could not find the provided ID in the database. This is problematic because I was potentially calling `dict(None)` in the return statement of two of our API endpoints. Now, I check to confirm that the return value is not `None` before processing.

Reviewed By: spurav

Differential Revision: D20452093

fbshipit-source-id: 0d86857fbf3bca2e84b847835d89e13a2da76271