Save network test iperf results to db
Summary:
This diff adds the ability to start saving the Iperf results from a network test in the database. So far we will save the raw message, src + dst nodes, and whether the src node was the server during the test. I've added the iperf throughput and link error data as well. In the next diff, I will add the Prometheus stats from during the test (mcs, snr, etc.).

This diff also adds logic to stop an execution if it takes longer than its estimated time duration. In addition, there is a lot of restructuring and cleanup

Reviewed By: pmccut

Differential Revision: D20427364

fbshipit-source-id: ec74c295ab6d49bcc748053229910fb2abe45f3e