Only normalize label string in Prometheus query
Summary:
There are cases where we want the inverse (1/x) or the complement (1-x) of a Prometheus metric. Instead of performing this analysis in `tglib` services, we can tell Prometheus to do this for us by passing the formula w/ operands directly as the metric name. i.e. `1 - pinger_lossRatio`

The current query creator normalizes the metric name along with the labels. This causes the operators to be changed into underscores too. This diff changes the formatting so that we only format the labels.

Reviewed By: hmehmood0

Differential Revision: D20625015

fbshipit-source-id: 47a1dc696540c875b599db1e3ccb4bd542d7a768