upload topology functionality
Summary: Enables users to upload a JSON file and automatically create topology based on the file. After much discussion on if this feature was on the same level as the other topology editing panels, we agreed for this version of TGNMS, all topology editing features should have the same look and feel and reside in the same place. For topology adding and editing details see design document: https://fb.quip.com/ujtbAlCrQza0

Reviewed By: pmccut

Differential Revision: D20285251

fbshipit-source-id: 06837b020e58a62647dae8e9df4e558989546bcb