Planned Site Template
Summary: The creation of sites, nodes and links in TGNMS is hard and not user friendly. The goal of this diff is to reduce the number of steps a user needs to manually create on site, with nodes, and links based on the type of site it is. Templating out basic site types allows us to do the majority of the work for the user and make creating a functioning network a better, quicker easier process. For more details see design document: https://fb.quip.com/ujtbAlCrQza0

Reviewed By: aclave1

Differential Revision: D20285237

fbshipit-source-id: 5b6334018852d0d93d9d942a1b5802a406449af3