fix websocket repeatedly reconnecting
Summary: a bug in the websocket library was causing heartbeats to fail, meaning websockets would repeatedly reconnect. working around the bug fixes the issue. I think the lib got downgraded but when I used docs from the older version, nothing changes. I think the docs may just be wrong or express-ws is decorating the websocket in a weird way.

Differential Revision: D20654605

fbshipit-source-id: bddba958c03d3ac4e79708e6e0709911ecc47ff7