Update prometheus client metric publishing
Summary:
The functionality for writing prometheus metrics was a little complicated so I refactored the code. The only service that used this functionality so far was the default routes service so I updated the code there to be congruent with the changes.

I also moved the default scrape interval for prometheus to `group_vars` so that tglib and prometheus could both use the same value.

Reviewed By: spurav

Differential Revision: D20673084

fbshipit-source-id: b056d81b2c81f491054d192a51e59bdca1e8e7d4