Default Grafana Dashboards for NMS (#404)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/404

Creates default dashboards that mirror the original nms dashboards. These include variables to select on `networkID` and `gatewayID` in the same way that the NMS dashboards do, except these allow the selection of multiple networks at once.

To deal with updates to the dashboard, they are overwritten on every sync and are not editable. If a user wants to modify the dashboard they can copy it and modify the copy.

**Update**: The problem with the first iteration was that the dependency was added under `devDependencies` and as such it wasn't included in the prod build.

Reviewed By: xjtian

Differential Revision: D20443215

fbshipit-source-id: 6b6aead9e8b75d9a193db8db528bd5bfe18636c5