Write starter producer/consumer model framework
Summary:
The analytics service is being migrated to run as a producer/consumer `tglib` service. This diff creates all of the necessary files for engineers to jump in and add new jobs to the service.

Eventually, I will write a script that can stub all of this autogeneration but that can be done later.

Reviewed By: spurav

Differential Revision: D20799990

fbshipit-source-id: 4e9ddb538098479dbe21d330ec47022bc3f9808e