Drop limit(1) in list_schedules
Summary:
The `list_schedules` function in the network test scheduler class was always only showing the latest schedule in the DB due to incorrect `ORDER BY` and `LIMIT 1` clauses.

This diff removes them so that all schedules will be returned if no ID is provided, and only the schedule with the corresponding ID is returned if an ID is provided.

Differential Revision: D20924021

fbshipit-source-id: c14ed7b06eb2dffa723625a79b12d5e067bf9129