Tidy up default_routes_service
Summary:
There were some issues with the previous default routes implementation. For example, updating the default routes leased a connection from MySQL for every node.

We will also now use the `start_time_ms` generated at the beginning of the job as the publish time for Prometheus and MySQL actions in order to be more consistent across our datastores.

Reviewed By: spurav

Differential Revision: D20715742

fbshipit-source-id: 3dfbc6fb5295ed3b8e27a057e0f52bb9d86a769d